Issue #44: expose fractional history quantities

+++ src/ProjectHopeLynden.Application/Inventory/InventoryCountHistoryListItem.cs
@@ -2,6 +2,6 @@ namespace ProjectHopeLynden.Application.Inventory;
 
 public sealed record InventoryCountHistoryListItem(
     DateTime CountedAtUtc,
-    int? PreviousQuantity,
-    int CountedQuantity,
-    int? QuantityChange);
+    double? PreviousQuantity,
+    double CountedQuantity,
+    double? QuantityChange);
